+++ engine/UnitTest.Core/src/core/collection/LinkedListTest.cpp
@@ -3,6 +3,7 @@
 
 #include "core/collection/LinkedList.h"
 #include "core/collection/Stack.h"
+#include "core/Concepts.h"
 #include "TestAllocator.h"
 
 #include <iostream>
@@ -27,7 +28,7 @@ namespace UnitTest {
                 TestAllocator a;
                 LinkedList<uint32_t, TestAllocator&> list(a);
 
-                Assert::AreEqual(list.Size, 0ull);
+                Assert::AreEqual(list.Size, (size_t)0);
                 Assert::AreEqual(list.Empty, true);
                 Assert::AreEqual(a.Allocations, 0);
             }
@@ -35,7 +36,7 @@ namespace UnitTest {
             {
                 LinkedList<TestClass> list;
 
-                Assert::AreEqual(list.Size, 0ull);
+                Assert::AreEqual(list.Size, (size_t)0);
                 Assert::AreEqual(list.Empty, true);
             }
             _____________________checkMemory();
@@ -45,7 +46,7 @@ namespace UnitTest {
             LinkedList<uint32_t, TestAllocator> testLinkedList = { 1, 2, 3 };
             LinkedList<uint32_t, TestAllocator> list = LinkedList<uint32_t, TestAllocator>(testLinkedList);
 
-            Assert::AreEqual(list.Size, 3ull);
+            Assert::AreEqual(list.Size, (size_t)3);
             Assert::AreEqual(list.Empty, false);
             Assert::AreEqual(list[0], 1u);
             Assert::AreEqual(list[1], 2u);
@@ -57,7 +58,7 @@ namespace UnitTest {
             LinkedList<uint32_t> testLinkedList = { 1, 2, 3 };
             LinkedList<uint32_t> list = std::move(testLinkedList);
 
-            Assert::AreEqual(list.Size, 3ull);
+            Assert::AreEqual(list.Size, (size_t)3);
             Assert::AreEqual(list.Empty, false);
             Assert::AreEqual(list[0], 1u);
             Assert::AreEqual(list[1], 2u);
@@ -68,7 +69,7 @@ namespace UnitTest {
         TEST_METHOD(LinkedList_constructorInitializer) {
             LinkedList<uint32_t> list = { 1, 2, 3 };
 
-            Assert::AreEqual(list.Size, 3ull);
+            Assert::AreEqual(list.Size, (size_t)3);
             Assert::AreEqual(list.Empty, false);
             Assert::AreEqual(list[0], 1u);
             Assert::AreEqual(list[1], 2u);
@@ -94,7 +95,7 @@ namespace UnitTest {
             LinkedList<uint32_t> list;
             list = std::move(testLinkedList);
 
-            Assert::AreEqual(list.Size, 3ull);
+            Assert::AreEqual(list.Size, (size_t)3);
             Assert::AreEqual(list.Empty, false);
             Assert::AreEqual(list[0], 1u);
             Assert::AreEqual(list[1], 2u);
@@ -107,7 +108,7 @@ namespace UnitTest {
                 LinkedList<uint32_t> list;
                 list = { 1, 2, 3 };
 
-                Assert::AreEqual(list.Size, 3ull);
+                Assert::AreEqual(list.Size, (size_t)3);
                 Assert::AreEqual(list.Empty, false);
                 Assert::AreEqual(list[0], 1u);
                 Assert::AreEqual(list[1], 2u);
@@ -122,7 +123,7 @@ namespace UnitTest {
                 LinkedList<uint32_t, TestAllocator&> list(a);
                 list.add(1);
 
-                Assert::AreEqual(list.Size, 1ull);
+                Assert::AreEqual(list.Size, (size_t)1);
                 Assert::AreEqual(list.Empty, false);
                 Assert::AreEqual(list[0], 1u);
                 Assert::AreEqual(a.Allocations, 1);
@@ -137,7 +138,7 @@ namespace UnitTest {
                 LinkedList<uint32_t, TestAllocator&> list(a);
                 list.addAll({ 1, 2, 3 });
 
-                Assert::AreEqual(list.Size, 3ull);
+                Assert::AreEqual(list.Size, (size_t)3);
                 Assert::AreEqual(list.Empty, false);
                 Assert::AreEqual(list[0], 1u);
                 Assert::AreEqual(list[1], 2u);
@@ -156,13 +157,13 @@ namespace UnitTest {
                 const LinkedList<uint32_t, TestAllocator&> testLinkedList = LinkedList<uint32_t, TestAllocator&>({ 1, 2, 3 }, a2);
                 list.addAll(testLinkedList);
 
-                Assert::AreEqual(testLinkedList.Size, 3ull);
+                Assert::AreEqual(testLinkedList.Size, (size_t)3);
                 Assert::AreEqual(testLinkedList.Empty, false);
                 Assert::AreEqual(testLinkedList[0], 1u);
                 Assert::AreEqual(testLinkedList[1], 2u);
                 Assert::AreEqual(testLinkedList[2], 3u);
 
-                Assert::AreEqual(list.Size, 3ull);
+                Assert::AreEqual(list.Size, (size_t)3);
                 Assert::AreEqual(list.Empty, false);
                 Assert::AreEqual(list[0], 1u);
                 Assert::AreEqual(list[1], 2u);
@@ -179,7 +180,7 @@ namespace UnitTest {
             LinkedList<uint32_t> list = { 1, 2, 3 };
             list.insert(1, 4);
 
-            Assert::AreEqual(list.Size, 4ull);
+            Assert::AreEqual(list.Size, (size_t)4);
             Assert::AreEqual(list.Empty, false);
             Assert::AreEqual(list[0], 1u);
             Assert::AreEqual(list[1], 4u);
@@ -192,7 +193,7 @@ namespace UnitTest {
             LinkedList<uint32_t> list = { 1, 2, 3 };
             list.set(1, 4);
 
-            Assert::AreEqual(list.Size, 3ull);
+            Assert::AreEqual(list.Size, (size_t)3);
             Assert::AreEqual(list.Empty, false);
             Assert::AreEqual(list[0], 1u);
             Assert::AreEqual(list[1], 4u);
@@ -205,7 +206,7 @@ namespace UnitTest {
                 LinkedList<uint32_t> list = { 1, 2, 3 };
                 auto val = list.get(1);
 
-                Assert::AreEqual(list.Size, 3ull);
+                Assert::AreEqual(list.Size, (size_t)3);
                 Assert::AreEqual(list.Empty, false);
                 Assert::AreEqual(list[0], 1u);
                 Assert::AreEqual(list[1], 2u);
@@ -224,7 +225,7 @@ namespace UnitTest {
             LinkedList<uint32_t> list = { 1, 2, 3 };
             uint32_t& val = list[1];
 
-            Assert::AreEqual(list.Size, 3ull);
+            Assert::AreEqual(list.Size, (size_t)3);
             Assert::AreEqual(list.Empty, false);
             Assert::AreEqual(list[0], 1u);
             Assert::AreEqual(list[1], 2u);
@@ -243,7 +244,7 @@ namespace UnitTest {
                 LinkedList<uint32_t> list = { 1, 2, 3 };
                 list.remove(2);
 
-                Assert::AreEqual(list.Size, 2ull);
+                Assert::AreEqual(list.Size, (size_t)2);
                 Assert::AreEqual(list.Empty, false);
                 Assert::AreEqual(list[0], 1u);
                 Assert::AreEqual(list[1], 3u);
@@ -252,7 +253,7 @@ namespace UnitTest {
                 LinkedList<uint32_t> list = { 1 };
                 list.remove(1);
 
-                Assert::AreEqual(list.Size, 0ull);
+                Assert::AreEqual(list.Size, (size_t)0);
                 Assert::AreEqual(list.Empty, true);
             }
             _____________________checkMemory();
@@ -263,7 +264,7 @@ namespace UnitTest {
                 LinkedList<uint32_t> list = { 1, 2, 3 };
                 list.removeAt(2);
 
-                Assert::AreEqual(list.Size, 2ull);
+                Assert::AreEqual(list.Size, (size_t)2);
                 Assert::AreEqual(list.Empty, false);
                 Assert::AreEqual(list[0], 1u);
                 Assert::AreEqual(list[1], 2u);
@@ -272,7 +273,7 @@ namespace UnitTest {
                 LinkedList<uint32_t> list = { 1 };
                 list.removeAt(0);
 
-                Assert::AreEqual(list.Size, 0ull);
+                Assert::AreEqual(list.Size, (size_t)0);
                 Assert::AreEqual(list.Empty, true);
             }
             _____________________checkMemory();
@@ -282,7 +283,7 @@ namespace UnitTest {
             LinkedList<uint32_t> list = { 1, 2, 3,4,5,6,7,8 };
             list.removeAll({ 2,4,1 });
 
-            Assert::AreEqual(list.Size, 5ull);
+            Assert::AreEqual(list.Size, (size_t)5);
             Assert::AreEqual(list.Empty, false);
 
             LinkedList<uint32_t> testLinkedList = { 3,5,6,7,8 };
@@ -296,7 +297,7 @@ namespace UnitTest {
                 std::vector<uint32_t> testVector = { 1, 2, 3 };
                 LinkedList<uint32_t> list = { 1, 2, 3 };
 
-                Assert::AreEqual(list.Size, 3ull);
+                Assert::AreEqual(list.Size, (size_t)3);
                 Assert::AreEqual(list.Empty, false);
 
                 size_t i = 0;
@@ -308,7 +309,7 @@ namespace UnitTest {
                 std::vector<uint32_t> testVector = { 1, 2, 3 };
                 const LinkedList<uint32_t> list = { 1, 2, 3 };
 
-                Assert::AreEqual(list.Size, 3ull);
+                Assert::AreEqual(list.Size, (size_t)3);
                 Assert::AreEqual(list.Empty, false);
 
                 size_t i = 0;
@@ -322,19 +323,19 @@ namespace UnitTest {
             {
                 LinkedList<uint32_t> list = { 1, 2, 3 };
 
-                Assert::AreEqual(list.Size, 3ull);
+                Assert::AreEqual(list.Size, (size_t)3);
                 Assert::AreEqual(list.Empty, false);
-                Assert::AreEqual(list.indexOf(2), 1ull);
-                Assert::AreEqual(list.indexOf(4), 3ull);
+                Assert::AreEqual(list.indexOf(2), (size_t)1);
+                Assert::AreEqual(list.indexOf(4), (size_t)3);
             }
 
             {
                 const LinkedList<uint32_t> list = { 1, 2, 3 };
 
-                Assert::AreEqual(list.Size, 3ull);
+                Assert::AreEqual(list.Size, (size_t)3);
                 Assert::AreEqual(list.Empty, false);
-                Assert::AreEqual(list.indexOf(2), 1ull);
-                Assert::AreEqual(list.indexOf(4), 3ull);
+                Assert::AreEqual(list.indexOf(2), (size_t)1);
+                Assert::AreEqual(list.indexOf(4), (size_t)3);
             }
             _____________________checkMemory();
         }
@@ -343,7 +344,7 @@ namespace UnitTest {
             {
                 LinkedList<uint32_t> list = { 1, 2, 3 };
 
-                Assert::AreEqual(list.Size, 3ull);
+                Assert::AreEqual(list.Size, (size_t)3);
                 Assert::AreEqual(list.Empty, false);
                 Assert::IsTrue(list.contains(1));
                 Assert::IsFalse(list.contains(4));
@@ -352,7 +353,7 @@ namespace UnitTest {
             {
                 const LinkedList<uint32_t> list = { 1, 2, 3 };
 
-                Assert::AreEqual(list.Size, 3ull);
+                Assert::AreEqual(list.Size, (size_t)3);
                 Assert::AreEqual(list.Empty, false);
                 Assert::IsTrue(list.contains(1));
                 Assert::IsFalse(list.contains(4));
@@ -364,7 +365,7 @@ namespace UnitTest {
             LinkedList<uint32_t> list = { 1, 2, 3 };
             LinkedList<uint32_t> list2 = { 1, 2, 3 };
 
-            Assert::AreEqual(list.Size, 3ull);
+            Assert::AreEqual(list.Size, (size_t)3);
             Assert::AreEqual(list.Empty, false);
             Assert::IsTrue(list == list2);
             _____________________checkMemory();
